refactor: apply review comments (#1564)
